+++ kvec.h
@@ -56,7 +56,7 @@ int main() {
 
 #define kvec_t(type)                                                           \
   struct {                                                                     \
-    size_t n, m;                                                               \
+    int n, m;                                                               \
     type *a;                                                                   \
   }
 #define kv_init(v) ((v).n = (v).m = 0, (v).a = 0)
@@ -94,10 +94,10 @@ int main() {
       ((v).a + ((v).n++))
 
 #define kv_a(type, v, i)                                                       \
-  (((v).m <= (size_t)(i)                                                       \
+  (((v).m <= (int)(i)                                                       \
         ? ((v).m = (v).n = (i) + 1, kv_roundup32((v).m),                       \
            (v).a = (type *)realloc((v).a, sizeof(type) * (v).m), 0)            \
-        : (v).n <= (size_t)(i) ? (v).n = (i) + 1 : 0),                         \
+        : (v).n <= (int)(i) ? (v).n = (i) + 1 : 0),                         \
    (v).a[(i)])
 
 #define kv_swap(type, v, i, j)                                                 \
